+++ src/TaskManagerBundle.php
@@ -16,7 +16,7 @@ final class TaskManagerBundle extends Bundle
 	/**
 	 * @inheritDoc
 	 */
-	public function getContainerExtension () : ?ExtensionInterface
+	public function getContainerExtension () : ExtensionInterface
 	{
 		return new ConfigurableBundleExtension(
 			$this,